+++ lib/yasmx/Value.cpp
@@ -332,7 +332,7 @@ Value::FinalizeScan(Expr& e, bool ssym_ok, int* pos)
 {
     ExprTerms& terms = e.getTerms();
     if (*pos < 0)
-        *pos += terms.size();
+        *pos += static_cast<int>(terms.size());
 
     ExprTerm& root = terms[*pos];
     if (!root.isOp())
